huge cleanup to how sotd data is sent to client
