+++ pages/sotd.tsx
@@ -1,51 +1,45 @@
 import DefaultLayout from "../components/layouts/DefaultLayout";
 import {GetServerSideProps, InferGetServerSidePropsType} from "next";
 import {PlaylistData} from "../types/database";
-import {revalidatePlaylist} from "../lib/util/revalidatePlaylist";
-import {PlaylistObjectTransformed} from "../lib/util/transformPlaylist";
-import useSWR from "swr";
 import clientPromise from "../lib/mongodb";
-import {sotdAPIResponse} from "../types/sotd";
 import {DateTime} from "luxon";
-import {HeardleGame} from "../components/Heardle/game";
+import {getSotd, getSotdResponse} from "../lib/getSotd";
+import dynamic from 'next/dynamic'
 
-const fetcher = (url: RequestInfo | URL) => fetch(url).then((res) => res.json());
+const DynamicHeardleGame = dynamic(() => import('../components/Heardle/game').then((mod) => mod.HeardleGame), {
+    loading: () => <p>Loading...</p>,
+    ssr: false,
+})
 
-const SongOfTheDay = ({ playlist }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
-    const { data, error } = useSWR<sotdAPIResponse>(
-        "/api/sotd",
-        fetcher
-    );
-
+const SongOfTheDay = ({ sotdData }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
     return (
         <div className="flex-grow flex flex-col m-3 space-y-3 text-white">
             <div className="flex flex-grow justify-center self-center w-full h-full">
                 {
-                    data?
-                        <HeardleGame playlist={playlist} sotd={data}/>
+                    sotdData?
+                        <DynamicHeardleGame playlist={sotdData.playlist} sotd={sotdData}/>
                         : null
                 }
             </div>
             <div className="fixed bottom-0 left-0 text-gray-400 text-[0.5rem]">
-                {`Snapshot id: ${(data?.snapshot_id === playlist.snapshot_id)? playlist.snapshot_id: "❌ Snapshot id mismatch"} | Date: ${DateTime.now().setZone("Asia/Seoul").setLocale("en-GB").toLocaleString({dateStyle: "long"})} ${(data)? `Day: ${data?.day}`: ""}`}
+                {`Snapshot id: ${sotdData.snapshot_id} | Date: ${DateTime.now().setZone("Asia/Seoul").setLocale("en-GB").toLocaleString({dateStyle: "long"})} Day: ${sotdData?.day}`}
             </div>
         </div>
     )
 }
 
-export const getServerSideProps: GetServerSideProps<{playlist: PlaylistObjectTransformed}> = async ({req, res}) => {
+export const getServerSideProps: GetServerSideProps<{sotdData:getSotdResponse}> = async ({res}) => {
     let client = await clientPromise
     let Playlists = client.db("CHGM").collection<PlaylistData>("playlists")
     await Playlists.createIndex({snapshot_id: 1}, {unique: true})
-    let {playlist} = await revalidatePlaylist(Playlists)
-
+    const sotdData = await getSotd()
     res.setHeader(
         'Cache-Control',
         'public, s-maxage=300, stale-while-revalidate=59'
     )
     return {
         props: {
-            playlist
+            sotdData
         },
     }
 }
