+++ types/sotd.ts
@@ -27,3 +27,4 @@ export interface sotdAPIResponse extends sotdGameData {
 }
 
 export type Nullable<T> = T | null | undefined;
+export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>
